Core/PacketIO: Fixed SMSG_SPELL_ENERGIZE_LOG structure
Port From (https://github.com/TrinityCore/TrinityCore/commit/ba90348de88f0355b559010f8a106527b094baa2)

+++ Source/Game/Networking/Packets/CombatLogPackets.cs
@@ -382,7 +382,7 @@ namespace Game.Networking.Packets
             _worldPacket.WritePackedGuid(CasterGUID);
 
             _worldPacket.WriteUInt32(SpellID);
-            _worldPacket.WriteUInt32((uint)Type);
+            _worldPacket.WriteInt8((sbyte)Type);
             _worldPacket.WriteInt32(Amount);
             _worldPacket.WriteInt32(OverEnergize);
